Stop form handler after read timeout, only if no data is left
Fixes #682 (fixes first fix)

+++ src/mod_lua.inl
@@ -390,7 +390,7 @@ lsp_var_reader(lua_State *L, void *ud, size_t *sz)
 	case 1:
 		/* Second call: forward variable name */
 		ret = reader->begin;
-		*sz = reader->len;
+		*sz = (size_t)reader->len;
 		reader->consumed += reader->len;
 		break;
 	case 2:
